Update Assignment2_Elin_Kile
Finished script. Reads csv file into data frame. In the csv there are two rows, one with samples and one with classification, 0 or 1. Class 0 is gamma distributed, class 1 is gaussian. The script makes one training test and one test set 80/20. Trains the model and tests it. Returns accuracy, and parameters for the distributions. Plots different histograms of the whole data set, the two classes, and the correctly classified and wrongly classified.